Clean up code structure
- Make dedicated TS files for code strings
- Check for euconsent-v2 cookie before triggering any clicker since it's
  a multivendor standard
