+++ src/indirectClickers/trustarc.ts
@@ -1,13 +1,9 @@
-import { clickWhenFound, hasEuConsentCookie } from "../common";
+import { clickWhenFound } from "../common";
 
 async function handle() {
-    if (hasEuConsentCookie(document.cookie)) {
-        console.debug('Assuming already opted out');
-    } else {
-        // This will trigger an iframe that is handled by the content script in the
-        // direct clicker folder
-        await clickWhenFound('.trustarc-manage-btn');
-    }
+    // This will trigger an iframe that is handled by the content script in the
+    // direct clicker folder
+    await clickWhenFound('.trustarc-manage-btn');
 }
 
 export default handle;
